games-demo/requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-logging
itables
streamlit-navigation-bar
streamlit-extras
streamlit-agraph
python-dotenv